requirements.txt
glfw==2.5.3
numpy==1.22.4
Pillow==9.1.1
PyGLM==2.5.7
PyOpenGL==3.1.6
